backend/requirements.txt
flasgger
flask_cors
eventlet
Flask-SocketIO==5.3.3
python-socketio==5.5.0
eventlet==0.33.3
flask-cors
flask-socketio
pandas
graphviz
requests
scipy